alertspage

+++ client/src/pages/AlertsPage.jsx
@@ -1,8 +1,14 @@
 import { useEffect } from 'react';
 import { useAlerts, useDocumentTitle } from '../hooks/index.js';
-import { toast } from 'react-hot-toast';
 
-// @@@ quitar error No se encontraron alertas declaradas para este usuario, y poner un sweetalert2 de confirmación al darle a eliminar alerta
+// Pop-ups
+import { toast } from 'react-hot-toast';
+import Swal from 'sweetalert2';
+
+// React icons
+import { FaTrashCan } from 'react-icons/fa6';
+
+// @@@ quitar error No se encontraron alertas declaradas para este usuario
 
 // ------------------------------------------
 const AlertsPage = () => {
@@ -24,8 +30,20 @@ const AlertsPage = () => {
 
   const handleDelete = async (alertId) => {
     try {
-      await deleteAlert(alertId);
-      toast.success('Alerta eliminada correctamente.');
+      // Usar SweetAlert2 para confirmar eliminación
+      const result = await Swal.fire({
+        title: '¿Estás seguro?',
+        text: '¡Esta acción no se puede deshacer!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Eliminar',
+        cancelButtonText: 'Cancelar',
+      });
+
+      if (result.isConfirmed) {
+        await deleteAlert(alertId);
+        toast.success('Alerta eliminada correctamente.');
+      }
     } catch (err) {
       toast.error(err.message);
     }
@@ -34,23 +52,39 @@ const AlertsPage = () => {
   const alertList = alerts.alerts || []; // Garantizamos un array vacío si no hay alertas.
 
   return (
-    <div>
-      <h2>Historial de alertas</h2>
+    <div className="h-fit px-5 py-10 text-darkgray">
+      <h2 className="mb-6 text-center text-3xl font-bold">
+        Historial de alertas
+      </h2>
       {alertList.length > 0 ? (
-        <ul>
+        <ul className="flex w-full max-w-xl space-y-4 px-4">
           {alertList.map((alert) => (
-            <li key={alert.id}>
+            <li
+              key={alert.id}
+              className="w-full rounded-lg bg-whitegray p-4 shadow-lg transition-shadow hover:shadow-xl"
+            >
               <div>
-                <h3>{alert.type}</h3>
-                <p>{`Valor: ${alert.value}`}</p>
-                <p>{`Fecha: ${new Date(alert.createdAt).toLocaleDateString()}`}</p>
-                <button onClick={() => handleDelete(alert.id)}>Eliminar</button>
+                <h3 className="text-xl font-semibold text-vibrantblue">
+                  {alert.type}
+                </h3>
+                <p className="text-gray">Valor: {alert.value}</p>
+                <p className="text-gray">
+                  Fecha: {new Date(alert.createdAt).toLocaleDateString()}
+                </p>
+                <button
+                  onClick={() => handleDelete(alert.id)}
+                  className="rounded-md bg-red p-2 text-whitegray transition-all hover:scale-105 hover:bg-warmyellow"
+                >
+                  <FaTrashCan />
+                </button>
               </div>
             </li>
           ))}
         </ul>
       ) : (
-        <p>No hay alertas disponibles.</p>
+        <p className="text-center text-2xl italic text-gray">
+          No hay alertas disponibles.
+        </p>
       )}
     </div>
   );
